+++ source/scene/Scene.cpp
@@ -9,9 +9,14 @@ https://github.com/mvxxx
 #include "mapManager/SelectionManager.hpp"
 
 Scene::Scene( const std::string& title, const sf::Vector2f& dimensions)
-	:viewSpeed( 10.f ), motionSensitivity( 50.f ), zoomSpeed( 0.05f )
+	:viewSpeed( 10.f ),
+	motionSensitivity( 50.f ),
+	zoomSpeed( 0.05f )
 {
-	window = std::shared_ptr<sf::RenderWindow>( new sf::RenderWindow( sf::VideoMode( dimensions.x, dimensions.y ), title) );
+	window = std::shared_ptr<sf::RenderWindow>(
+		new sf::RenderWindow( 
+			sf::VideoMode( static_cast<unsigned int>(dimensions.x), static_cast<unsigned int>(dimensions.y) ), 
+			title) );
 
 	views[viewType_t::DEFAULT] = std::make_shared<sf::View>();
 	views[viewType_t::UI] = std::make_shared<sf::View>();
